+++ example/lib/example5_themes_playground/theme/flex_theme_dark.dart
@@ -382,6 +382,9 @@ FlexColorScheme flexColorSchemeDark(ThemeController controller, Color source) {
             navigationBarElevation: controller.navBarElevation,
             navigationBarHeight: controller.navBarHeight,
             navigationBarLabelBehavior: controller.navBarLabelBehavior,
+            adaptiveRemoveNavigationBarTint: controller
+                .adaptiveRemoveNavigationBarTintDark
+                ?.setting(controller.fakeIsWeb),
             // NavigationRail settings.
             navigationRailSelectedLabelSchemeColor:
                 controller.navRailSelectedLabelSchemeColor,