+++ example/lib/example5_themes_playground/theme/flex_theme_light.dart
@@ -446,6 +446,9 @@ FlexColorScheme flexColorSchemeLight(ThemeController controller, Color source) {
             navigationBarElevation: controller.navBarElevation,
             navigationBarHeight: controller.navBarHeight,
             navigationBarLabelBehavior: controller.navBarLabelBehavior,
+            adaptiveRemoveNavigationBarTint: controller
+                .adaptiveRemoveNavigationBarTintLight
+                ?.setting(controller.fakeIsWeb),
             // NavigationRail settings.
             navigationRailSelectedLabelSchemeColor:
                 controller.navRailSelectedLabelSchemeColor,